python/transliteration/requirements.txt
ai4bharat-transliteration==1.1.3
alabaster==1.0.0
antlr4-python3-runtime==4.8
babel==2.17.0
bitarray==3.7.1
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Cython==3.1.3
docutils==0.21.2
-e git+ssh://git@github.com-personal/facebookresearch/fairseq.git@82b58777e16932259e575a504e24ceb12b08cec5#egg=fairseq
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
fsspec==2025.9.0
gevent==25.8.2
greenlet==3.2.4
hydra-core==1.0.7
idna==3.10
imagesize==1.4.1
indic-nlp-library==0.92
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.1
MarkupSafe==3.0.2
mock==5.2.0
Morfessor==2.0.6
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
omegaconf==2.0.6
packaging==25.0
pandas==2.3.2
portalocker==3.2.0
progressbar2==4.5.0
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.23
pydload==1.0.9
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-utils==3.9.1
pytz==2025.2
pywin32==311
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
roman-numerals-py==3.1.0
sacrebleu==2.5.1
sacremoses==0.1.1
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinx-argparse==0.5.2
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
torch==2.8.0
torchaudio==2.8.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urduhack==0.1.4
urllib3==2.5.0
Werkzeug==3.1.3
zope.event==6.0
zope.interface==8.0
